Project restructuring + minor bug fixes
