+++ src/interpreter/exceptions/break_exception.py
@@ -1,0 +1,3 @@
+class BreakException(Exception):
+    """Exception raised when a break statement is encountered"""
+    pass
