+++ src/interpreter/exceptions/infinte_loop_exception.py
@@ -1,0 +1,3 @@
+class InfiniteLoopException(Exception):
+    """Exception raised when an infinite loop is detected"""
+    pass
