+++ tests/unit/images.py
@@ -1,40 +1,9 @@
-import tempfile
-import os
 import io
 import unittest.mock
-from functools import wraps
-
 import exiftool
-
-from image_metadata import metadata_prefix
-from main import parser
-
-
-def get_output(code, mock_stdout):
-    ast = parser.parse(code, tracking=True)
-    ast.eval()
-    output = mock_stdout.getvalue().strip('\n')
-    return output
-
-
-def with_tempfile(extension):
-    def decorator(test_func):
-        @wraps(test_func)
-        def wrapper(self, *args, **kwargs):
-            with tempfile.NamedTemporaryFile(suffix=extension, delete=False) as tmp:
-                temp_path = tmp.name
-                try:
-                    if extension in ('png', 'jpg', 'jpeg', 'gif'):
-                        tmp.flush()
-                        os.system(f"cp ../../test.{extension} {temp_path}")
-                    return test_func(self, temp_path, *args, **kwargs)
-                finally:
-                    try:
-                        os.unlink(temp_path)
-                    except OSError:
-                        pass
-        return wrapper
-    return decorator
+from src.utils.image_metadata import metadata_prefix
+from utils.get_output import *
+from utils.with_tempfile import *
 
 
 class ImagesTestCase(unittest.TestCase):
@@ -51,11 +20,13 @@ class ImagesTestCase(unittest.TestCase):
             self.assertEqual(content in output, True)
 
     @with_tempfile('png')
-    def test_image_set_title_no_prefix(self, temp_file):
-        field = "Title"
+    def test_image_set_comment_no_prefix(self, temp_file):
+        field = "Comment"
         code = f'file f = load("{temp_file}")\nset(f,"{field}","a")'
         ast = parser.parse(code, tracking=True)
         ast.eval()
+
+        os.system(f"echo {temp_file}")
 
         with exiftool.ExifToolHelper() as et:
             metadata = et.get_metadata(temp_file)[0]
@@ -63,8 +34,8 @@ class ImagesTestCase(unittest.TestCase):
             self.assertEqual(metadata[key], "a")
 
     @with_tempfile('png')
-    def test_image_set_title_with_prefix(self, temp_file):
-        field = "Title"
+    def test_image_set_comment_with_prefix(self, temp_file):
+        field = "Comment"
         key = f"{metadata_prefix(field)}:{field}"
         code = f'file f = load("{temp_file}")\nset(f,"{key}","a")'
         print(code)
