+++ tests/utils/get_output.py
@@ -1,0 +1,8 @@
+from main import parser
+
+
+def get_output(code, mock_stdout):
+    ast = parser.parse(code, tracking=True)
+    ast.eval()
+    output = mock_stdout.getvalue().strip('\n')
+    return output
